refactor(hir): Split out signature.rs
