+++ src/hir/class_dict/mod.rs
@@ -35,7 +35,7 @@ impl ClassDict {
                           clsname: &ClassFullname,
                           defs: &[ast::Definition]) -> Vec<MethodParam> {
         if let Some(ast::Definition::InstanceMethodDefinition { sig, .. }) = defs.iter().find(|d| d.is_initializer()) {
-            hir::convert_params(&sig.params)
+            hir::signature::convert_params(&sig.params)
         }
         else {
             let (sig, _found_cls) =
